docs(Table): editable table sample updated
