--- packages/core/src/Table/stories/TableEditable/TableEditable.stories.mdx
@@ -1,13 +1,0 @@
-import { Meta, Canvas, Story } from "@storybook/addon-docs";
-import { EditableStory } from "./TableEditableSamples";
-
-<Meta title="Visualizations/Table/Editable Table" />
-
-# Editable Table
-
-This sample showcases an editable table where new rows can be added and rows can be edited and deleted.
-[`HvTableSection`](./?path=/docs/visualizations-table-table-section--docs) was used to use the new visual changes from Design System.
-
-<Canvas>
-  <Story name="Editable Table" story={EditableStory} />
-</Canvas>
