+++ packages/core/src/Table/stories/TableRenderers/DropdownColumnRenderer.tsx
@@ -118,7 +118,7 @@ export const DropdownColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
