+++ packages/core/src/Table/stories/TableRenderers/ExpandColumnRenderer.tsx
@@ -126,7 +126,7 @@ export const ExpandColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={50} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
