+++ packages/core/src/Table/stories/TableRenderers/NumberColumnRenderer.tsx
@@ -96,7 +96,7 @@ export const NumberColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={50} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
