+++ packages/core/src/Table/stories/TableRenderers/ProgressColumnRenderer.tsx
@@ -105,7 +105,7 @@ export const ProgressColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
