+++ packages/core/src/Table/stories/TableRenderers/SwitchColumnRenderer.tsx
@@ -112,7 +112,7 @@ export const SwitchColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={50} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
