+++ packages/core/src/Table/stories/TableRenderers/TagColumnRenderer.tsx
@@ -100,7 +100,7 @@ export const TagColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={50} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
