+++ packages/core/src/Table/stories/TableRenderers/TextColumnRenderer.tsx
@@ -96,7 +96,7 @@ export const TextColumnRenderer = () => {
             </HvTableRow>
           </HvTableHead>
           <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow /> : rowRenderer(page)}
+            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
           </HvTableBody>
         </HvTable>
       </HvTableContainer>
